requirements.txt
flask
opencv-python
mediapipe==0.10.11
numpy
pillow
arabic-reshaper
python-bidi
scikit-learn